refactor(dashboard): finished refactoring of dashboard
