+++ src/app/not-found.tsx
@@ -6,14 +6,14 @@ export default function NotFoundPage() {
       <div className="flex flex-col items-center space-y-5">
         <h1 className="text-8xl font-bold">404</h1>
         <h3 className="text-2xl font-semibold">
-          Sorry, we couldn't find this page!
+          Sorry, we couldn&apos;t find this page!
         </h3>
         <h4 className="text-xl">
           Maybe try again in{" "}
           <a
             target="_blank"
             href="/dashboard"
-            className="text-blue-700 hover:underline"
+            className="bg-white text-blue-700 hover:underline"
           >
             Dashboard
           </a>{" "}
