--- src/app/test_dashboard/page.tsx
@@ -1,5 +1,0 @@
-function DashboardPage() {
-  return null;
-}
-
-export default DashboardPage;
